requirements.txt
# requirements.txt
flask==2.3.3
supabase==1.0.3
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.143
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.24.3
opencv-python==4.8.0.76